feat(validators,-exceptions,-responses): Param Validator Objects, New Exception Object
BREAKING CHANGE: Completed the creation of the validator classes to validate requests. Also, added the ValidatorError exception class to be registered as part of the exceptions to be expected. Please refer to *register_error_handlers()* in the *create_app.py* file.

TODO: Implement authorization module and add DAOs from Database.